requirementstwo.txt
python-dotenv==1.0.1
gradio==4.44.1
pinecone==5.4.2
langchain==0.3.15
langchain-openai==0.3.2
langchain-pinecone==0.2.2
tiktoken==0.8.0
ory-kratos-client==1.0.0
requests==2.31.0